remove logic for setting changeType; separate out getFeatureMap function, continues #115
